file structure error fix
